# openapi.yaml
openapi: 3.0.3
info:
  title: Lightning Engine API
  description: |
    REST API for the Lightning Engine ECS game server.

    ## Overview

    The engine uses a tick-based simulation model:
    1. **Commands** are queued via POST requests or WebSocket
    2. **Tick** advances process commands and run all module systems
    3. **Snapshots** return the current ECS state

    ## WebSocket Endpoints

    ### Snapshot Streaming
    Connect via WebSocket at `ws://localhost:8080/ws/containers/{containerId}/snapshots/{matchId}` for real-time updates.

    ### Command WebSocket (Recommended for high-performance)
    Connect via WebSocket at `ws://localhost:8080/containers/{containerId}/commands?token=xxx` to submit commands using Protocol Buffers.

    **Authentication:** JWT token passed as query parameter (must have `admin` or `command_manager` role).

    **Protocol:** Binary Protocol Buffer messages (`CommandRequest`/`CommandResponse`).

    See `api-proto/src/main/proto/command.proto` for message definitions.

  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/modules/{moduleName}:
    get:
      tags: [Modules]
      summary: Get module details
      operationId: getModule
      parameters:
        - name: moduleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Module details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleResponse'
        '404':
          description: Module not found
    delete:
      tags: [Modules]
      summary: Uninstall module
      operationId: uninstallModule
      parameters:
        - name: moduleName
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Module uninstalled
        '404':
          description: Module not found
...
  /api/resources/{resourceId}:
    get:
      tags: [Resources]
      summary: Get resource metadata
      operationId: getResource
      parameters:
        - $ref: '#/components/parameters/resourceId'
      responses:
        '200':
          description: Resource metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '404':
          description: Resource not found
    delete:
      tags: [Resources]
      summary: Delete resource
      operationId: deleteResource
      parameters:
        - $ref: '#/components/parameters/resourceId'
      responses:
        '204':
          description: Resource deleted
        '404':
          description: Resource not found
...
  /api/modules/upload:
    post:
      tags: [Control Plane - Modules]
      summary: Upload a module
      description: Upload a module JAR to the registry
      operationId: uploadClusterModule
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                description:
                  type: string
              required:
                - file
      responses:
        '201':
          description: Module uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CPModuleResponse'
        '400':
          description: Invalid file
  /api/modules/{name}:
    get:
      tags: [Control Plane - Modules]
      summary: List module versions
      description: Get all versions of a module
      operationId: listModuleVersions
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Module versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CPModuleResponse'
  /api/modules/{name}/{version}:
    get:
      tags: [Control Plane - Modules]
      summary: Get module metadata
      operationId: getModuleMetadata
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Module metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CPModuleResponse'
        '404':
          description: Module not found
    delete:
      tags: [Control Plane - Modules]
      summary: Delete a module version
      operationId: deleteModuleVersion
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Module deleted
        '404':
          description: Module not found
...
components:
  parameters:
...
    resourceId:
      name: resourceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    ModuleResponse:
      type: object
      properties:
        name:
          type: string
          description: Module name (e.g., "SpawnModule")
        flagComponentName:
          type: string
          nullable: true
          description: Name of the module's flag component
        enabledMatches:
          type: integer
          description: Number of matches with this module enabled
    ResourceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
          enum: [TEXTURE]
        size:
          type: integer
          description: Size in bytes
...
    CPModuleResponse:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        description:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        checksum:
          type: string
        uploadedAt:
          type: string
          format: date-time
        uploadedBy:
          type: string
        downloadUrl:
          type: string
...